requirements.txt
gunicorn==20.0.4
starlette==0.14.1
uvicorn==0.12.3
tortoise-orm==0.16.18
asyncpg==0.21.0
pydantic==1.7.2
